% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_primer3.R
\name{design_primer}
\alias{design_primer}
\title{Design Primers}
\usage{
design_primer(fastafile, ...)
}
\arguments{
\item{fastafile}{sequence in fasta format}

\item{...}{other parameters passed to Primer3}
}
\value{
primers
}
\description{
Design Primers
}
\examples{
  fastafile = system.file("sequence.fa", package = "rPrimer3")
  design_primer(fastafile)
}
