% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_primer.R
\name{design_primer_from_file}
\alias{design_primer_from_file}
\title{Design Primers}
\usage{
design_primer_from_file(
  fastafile,
  filter_description_pattern = NULL,
  parts = 3,
  part_length = 500,
  ...
)
}
\arguments{
\item{fastafile}{sequence in fasta format}

\item{filter_description_pattern}{a string pattern}

\item{parts}{separate long template to parts}

\item{part_length}{length of part}

\item{...}{other parameters passed to Primer3}
}
\value{
primers
}
\description{
Design Primers
}
\examples{
  fastafile = system.file("sequence.fa", package = "rPrimer3")
  design_primer_from_file(fastafile)
}
