% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_fetch_seq.R
\name{ncbi_fetch_seq}
\alias{ncbi_fetch_seq}
\title{Fetch sequence from NCBI}
\usage{
ncbi_fetch_seq(
  term,
  db_from = "assembly",
  db_to = "nuccore",
  rettype = c("fasta", "gb"),
  retmode = "text",
  file = paste(term, rettype, sep = ".")
)
}
\arguments{
\item{term}{search term}

\item{db_from}{search db}

\item{db_to}{default: nuccore}

\item{rettype}{default: fasta}

\item{file}{file to write sequence}
}
\value{
NULL, if write successfully
}
\description{
Fetch sequence from NCBI
}
\examples{
\dontrun{
  ncbi_fetch_seq("GCA_023823755.1")
}
}
