% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_boulder.R
\name{read_boulder}
\alias{read_boulder}
\title{Read boulder to list}
\usage{
read_boulder(x)
}
\arguments{
\item{x}{text vector in Boulder format}
}
\value{
a list
}
\description{
Read boulder to list
}
\examples{
  read_boulder(x = c("A=1","B=2"))
}
