% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_boulder.R
\name{write_boulder}
\alias{write_boulder}
\title{Write a list config to Boulder-IO format}
\usage{
write_boulder(x, file = tempfile())
}
\arguments{
\item{x}{a named list}

\item{file}{output file}
}
\description{
Write a list config to Boulder-IO format
}
