% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_primer.R
\name{design_primer}
\alias{design_primer}
\title{Design Primer from a `SeqFastadna` object}
\usage{
design_primer(fasta, parse = TRUE, ...)
}
\arguments{
\item{fasta}{a `SeqFastadna` object}

\item{parse}{parse result}

\item{...}{other parameters passed to Primer3}
}
\value{
primer output
}
\description{
Design Primer from a `SeqFastadna` object
}
\examples{
fastafile = system.file("sequence.fa", package = "rPrimer3")
fasta = seqinr::read.fasta(file = fastafile, seqtype = "DNA", as.string = TRUE)
design_primer(fasta[[1]])
}
